hitl-contributor/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hitl-contributor"
version = "1.0.0"
description = "Human-in-the-Loop AI agent for open source contributions"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    "PyGithub>=2.3.0",
    "gitpython>=3.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "docker>=7.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "httpx>=0.27.0",   # for FastAPI TestClient
]

[project.scripts]
hitl-run   = "main:run_agent"
hitl-web   = "web_server:main"
hitl-sched = "scheduler:main"

# ── Ruff (lint + format) ──────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]   # line length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ── pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths   = ["tests"]
asyncio_mode = "auto"
addopts     = "-v --tb=short"

# ── mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = true
